Use SINT 'host' configuration option
Recently, the Smack Integration Test Framework got a new configuration option, named 'host'. This allows the test to be configured with an IP address or DNS name of the XMPP service to run the tests on.

This commit renames the old 'ipAddress' configuration option to 'host', to reflect the changes in each of the corresponding projects, to improve consistency in nomenclature.